[crypto] move knet_handle_get_crypto_list to knet_get_crypto_list
now matches the same API as get_transport_list
