+++ libknet/crypto.c
@@ -184,43 +184,27 @@ void crypto_fini(
 	return;
 }
 
-int knet_handle_get_crypto_list(knet_handle_t knet_h, const char **crypto_names, size_t *num_names)
+int knet_get_crypto_list(struct knet_crypto_info *crypto_list, size_t *crypto_list_entries)
 {
-	int savederrno = 0;
 	int err = 0;
 	int idx = 0;
 	int outidx = 0;
 
-	if (!knet_h) {
+	if (!crypto_list_entries) {
 		errno = EINVAL;
-		return -1;
-	}
-
-	if (!num_names) {
-		errno = EINVAL;
-		return -1;
-	}
-
-	savederrno = pthread_rwlock_rdlock(&knet_h->global_rwlock);
-	if (savederrno) {
-		log_err(knet_h, KNET_SUB_HANDLE, "Unable to get write lock: %s",
-			strerror(savederrno));
-		errno = savederrno;
 		return -1;
 	}
 
 	while (crypto_modules_cmds[idx].model_name != NULL) {
 		if (crypto_modules_cmds[idx].built_in) {
-			if (crypto_names) {
-				crypto_names[outidx] = crypto_modules_cmds[idx].model_name;
+			if (crypto_list) {
+				crypto_list[outidx].name = crypto_modules_cmds[idx].model_name;
 			}
 			outidx++;
 		}
 		idx++;
 	}
-	*num_names = outidx;
+	*crypto_list_entries = outidx;
 
-	pthread_rwlock_unlock(&knet_h->global_rwlock);
-	errno = savederrno;
 	return err;
 }
